# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/episodes/{id}": {
      "get": {
        "operationId": "EpisodesController_getEpisode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "episodes"
        ]
      },
      "put": {
        "operationId": "EpisodesController_updateEpisode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "episodes"
        ]
      }
    },
    "/api/v1/episodes/{id}/characters": {
      "get": {
        "operationId": "EpisodesController_listEpisodeCharacters",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "episodes"
        ]
      }
    },
    "/api/v1/episodes/{id}/pipeline-status": {
      "get": {
        "operationId": "EpisodesController_getEpisodePipelineStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "episodes"
        ]
      }
    },
...
    "/api/v1/episodes/{id}/scenes": {
      "get": {
        "operationId": "ScenesController_listEpisodeScenes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "scenes"
        ]
      }
    },
...
    "/api/v1/scenes/{id}": {
      "put": {
        "operationId": "ScenesController_updateScene",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "scenes"
        ]
      },
      "delete": {
        "operationId": "ScenesController_deleteScene",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "scenes"
        ]
      }
    },
...
  },
  "info": {
    "title": "XIAOCHUANG Backend API",
    "description": "V2.0 unified backend for web and admin clients",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}